+++ aioxmpp/xso/model.py
@@ -688,7 +688,7 @@ class Collector(_PropBase):
 
     def to_sax(self, instance, dest):
         for node in self.__get__(instance, type(instance)):
-            lxml.sax.saxify(node, StartDocumentFilter(dest))
+            lxml.sax.saxify(node, _CollectorContentHandlerFilter(dest))
 
 
 class Attr(Text):
@@ -2568,7 +2568,7 @@ def events_to_sax(events, dest):
             dest.characters(ev_args[0])
 
 
-class StartDocumentFilter(xml.sax.handler.ContentHandler):
+class _CollectorContentHandlerFilter(xml.sax.handler.ContentHandler):
     def __init__(self, receiver):
         super().__init__()
         self.__receiver = receiver
@@ -2587,12 +2587,6 @@ class StartDocumentFilter(xml.sax.handler.ContentHandler):
 
     def endElementNS(self, name, qname):
         self.__receiver.endElementNS(name, qname)
-
-    def startPrefixMapping(self, prefix, uri):
-        self.__receiver.startPrefixMapping(prefix, uri)
-
-    def endPrefixMapping(self, prefix):
-        self.__receiver.endPrefixMapping(prefix)
 
     def characters(self, content):
         self.__receiver.characters(content)